migrated NetworkLimits struct to rpc
